Add rule for gracefully handling button mappings
This one was tricky; we needed to allow strings to have full control
(and, in fact, we may well need to get fancier in the future depending
on the syntactical allowances of what button defs can do) within this
context, but we obviously don't want to duplicate their functionality.
Fortunately, we don't have to capture the whole line with a single
regex; instead, we can just match certain words and include the strings
context.

I suppose it's worth noting that there is the possibility that someone
writes multiple strings or multiple "means", but that's kind of on them.
Maybe we'll catch that later, I really don't know.